Compile entire source
Compiling just the main file works if the imports form a tree containing
the entire list of source files (thereby compiling all the files).

In general, this may not be true, and we want to compile all the files.
So do so explicitly.

We no longer need MAIN_FILE, so get rid of it.